Move java files to vw980 directory (as per new package structure)
