+++ java/src/main/java/vw980/vowpalWabbit/VW.java
