+++ java/src/main/java/vw980/vowpalWabbit/learner/VWBase.java
