+++ java/src/main/java/vw980/vowpalWabbit/learner/VWMulticlassLearner.java
