+++ java/src/main/java/vw980/vowpalWabbit/responses/ActionProbs.java
