+++ java/src/test/java/vw980/vowpalWabbit/ClosingTest.java
