+++ java/src/test/java/vw980/vowpalWabbit/VWTestHelper.java
