+++ java/src/test/java/vw980/vowpalWabbit/learner/VWScalarLearnerTest.java
